+++ cmd/main.go
@@ -170,6 +170,13 @@ func main() {
 		if err != nil {
 			log.Fatalf("Unable to connect to telescope ID %d: %s", telescopeId, err.Error())
 		}
+		connected, err = scope.GetConnected()
+		if err != nil {
+			log.Fatalf("Unable to determine status of telescope: %s", err.Error())
+		}
+		if !connected {
+			log.Fatalf("Telescope is not connected to ASCOM Remote")
+		}
 	}
 
 	name, err := scope.GetName()